+++ source/ProcessManager.Orchestrations/InternalProcesses/MigrateCalculationsFromWholesale/Wholesale/Model/CalculationExecutionState.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.WholesaleMigration.Wholesale.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.MigrateCalculationsFromWholesale.Wholesale.Model;
 
 /// <summary>
 /// IMPORTANT: Do not change numeric values as it'll affect persistence or communication.
