+++ source/ProcessManager.Orchestrations/InternalProcesses/MigrateCalculationsFromWholesale/Wholesale/Model/CalculationOrchestrationState.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.WholesaleMigration.Wholesale.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.MigrateCalculationsFromWholesale.Wholesale.Model;
 
 /// <summary>
 /// Used to represent the state of a calculation orchestration.
