+++ source/ProcessManager.Orchestrations/InternalProcesses/MigrateCalculationsFromWholesale/Wholesale/Model/QuantityUnit.cs
@@ -12,11 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.WholesaleMigration.Wholesale.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.InternalProcesses.MigrateCalculationsFromWholesale.Wholesale.Model;
 
-public sealed record OrchestrationInstanceId(string Id)
+/// <summary>
+/// The unit of measurement for the quantity.
+/// </summary>
+public enum QuantityUnit
 {
-    public string Id { get; } = !string.IsNullOrWhiteSpace(Id)
-        ? Id
-        : throw new ArgumentException($"Id was null or whitespace (value: \"{Id}\")", nameof(Id));
+    /// <summary>
+    /// The quantity unit is Kilo Watt Hour.
+    /// Code: H87
+    /// </summary>
+    Kwh,
+
+    /// <summary>
+    /// The quantity unit is pieces.
+    /// The unit is used for subscriptions and fees that are associated with the metering point.
+    /// </summary>
+    Pieces,
 }
