+++ source/ProcessManager.Orchestrations/ProcessManager.Orchestrations.csproj
@@ -65,8 +65,4 @@
     </Protobuf>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="InternalProcesses\WholesaleMigration\V1\" />
-  </ItemGroup>
-
 </Project>
